+++ examples/bad/test_aaa05.py
@@ -1,7 +1,20 @@
 import pytest
 
 
-def test_arrange():
+@pytest.fixture
+def fixture_a() -> str:
+    return 'A pointless fixture'
+
+
+@pytest.fixture
+def fixture_b() -> str:
+    return 'Another pointless fixture'
+
+
+# --- TESTS ---
+
+
+def test_arrange() -> None:
     """
     Blank line in Arrange Block
     """
@@ -9,27 +22,27 @@ def test_arrange():
 
     y = 4
 
-    result = x ** 2 + y ** 2
+    result = x**2 + y**2
 
     assert result == 25
 
 
-def test_act():
+def test_act() -> None:
     """
     Blank line in Act Block
     """
-    nothing = None
+    empty: list = []
 
-    with pytest.raises(AttributeError):
+    with pytest.raises(IndexError):
 
-        nothing.get_something()
+        empty[0]
 
 
-def test_assert():
+def test_assert() -> None:
     """
     Blank line in Assert Block
     """
-    result = list()
+    result: list = list()
 
     assert not result
 
@@ -37,10 +50,10 @@ def test_assert():
 
 
 def test_all(
-    fixture_a,
+    fixture_a: str,
 
-    fixture_b,
-):
+    fixture_b: str,
+) -> None:
     """
     Blank lines everywhere